消除Warning

+++ LeetCode/Problems/Easy/MST0507.cs
@@ -6,7 +6,7 @@
     {
         public int ExchangeBits(int num)
         {
-            return (int) (((num & 0x55555555) << 1) | ((num & 0xaaaaaaaa) >> 1));
+            return (int)((uint)((num & 0x55555555) << 1) | ((num & 0xaaaaaaaa) >> 1));
         }
     }
 }
